backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
cffi==1.17.1
cryptography==43.0.1
dj-database-url==3.0.1
Django==5.1.2
django-cors-headers==4.5.0
django-filter==24.3
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.0
gunicorn==23.0.0
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
redis==5.1.1
rsa==4.9
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==1.26.20
whitenoise==6.8.2
africastalking==1.2.8
celery[redis]==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
